Add marker to show which file is being tested. Add gaurd if temp file disappears

+++ test/test_legal_markdown_to_markdown.rb
@@ -29,11 +29,12 @@ class TestLegalMarkdownToMarkdown < Test::Unit::TestCase
   end
 
   def destroy_temp ( temp_file )
-    File.delete temp_file
+    File.delete temp_file if File::exists?(temp_file)
   end
 
   def test_files
     @lmdfiles.each do | lmd_file |
+      puts "Testing => #{lmd_file}"
       temp_file = create_temp
       benchmark_file = File.basename(lmd_file, ".lmd") + ".md"
       LegalToMarkdown.new( [ lmd_file, temp_file ] )
